Refactor Maps class to Maps and MapsBase
